# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Mandate Hackathon API",
    "version": "0.1.0",
    "description": "Contract-first prototype: one trusted FastAPI transaction service, scoped agent access, independently retained audit checkpoints and a bounded verification component. HKD only. Core wallet rails are simulated. See README.md for normative transaction, delegation and evidence requirements. Optional commerce extension exposes local sandbox checkout separately from the local payment simulator; merchant orders remain synthetic."
  },
  "servers": [
    {
      "url": "http://localhost:8000/api/v1",
      "description": "Local API; expose only scoped ingress when needed."
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/mandates/draft": {
      "post": {
        "operationId": "draftMandate",
        "tags": [
          "mandate"
        ],
        "summary": "Interpret user text into an inactive draft",
        "description": "Model output is a proposal. Unresolved ambiguities must be confirmed through the final explicit Policy.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-roles": [
          "user"
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            },
            "description": "Stable UUID recommended. Scoped by authenticated actor + operation. Same key and same semantic request returns the original outcome; changed request returns 409."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Resource created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated actor has insufficient authority.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found or not visible to caller.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Idempotency/payload conflict or illegal state transition.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Malformed or invalid request; distinct from a valid purchase refused by policy.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Temporary unavailable service or database contention; no payment performed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/agent-runs": {
      "post": {
        "operationId": "startAgentRun",
        "tags": [
          "agent"
        ],
        "summary": "Start a shopping run",
        "description": "Returns promptly; polling reads progress. Auto-purchase still uses the same wallet APIs and confirmed mandate.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-roles": [
          "user"
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            },
            "description": "Stable UUID recommended. Scoped by authenticated actor + operation. Same key and same semantic request returns the original outcome; changed request returns 409."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentRunRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful request; business refusal may be returned inside the decision object.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRun"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated actor has insufficient authority.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found or not visible to caller.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Idempotency/payload conflict or illegal state transition.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Malformed or invalid request; distinct from a valid purchase refused by policy.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Temporary unavailable service or database contention; no payment performed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/agent-runs/{run_id}": {
      "get": {
        "operationId": "getAgentRun",
        "tags": [
          "agent"
        ],
        "summary": "Read run progress",
        "description": "",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-roles": [
          "user"
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful request; business refusal may be returned inside the decision object.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRun"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated actor has insufficient authority.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found or not visible to caller.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Idempotency/payload conflict or illegal state transition.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Malformed or invalid request; distinct from a valid purchase refused by policy.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Temporary unavailable service or database contention; no payment performed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Server resolves an opaque configured credential to actor ID, role and authorized resources. Do not accept role/owner from request headers or body. Separate demo user and agent secrets; do not expose user secret to the agent."
      }
    },
    "schemas": {
      "Currency": {
        "type": "string",
        "enum": [
          "HKD"
        ]
      },
...
      "Category": {
        "type": "string",
        "enum": [
          "produce",
          "dairy",
          "eggs",
          "meat",
          "seafood",
          "bakery",
          "pantry",
          "beverage_non_alcoholic",
          "alcohol",
          "household",
          "unknown"
        ]
      },
...
      "PeriodLimit": {
        "properties": {
          "period": {
            "type": "string",
            "enum": [
              "calendar_week",
              "calendar_month"
            ],
            "title": "Period"
          },
          "limit_minor": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Limit Minor"
          },
          "timezone": {
            "type": "string",
            "const": "Asia/Hong_Kong",
            "title": "Timezone"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "period",
          "limit_minor",
          "timezone"
        ],
        "title": "PeriodLimit"
      },
      "Policy": {
        "properties": {
          "currency": {
            "type": "string",
            "const": "HKD",
            "title": "Currency"
          },
          "per_order_limit_minor": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Per Order Limit Minor"
          },
          "period_limits": {
            "items": {
              "$ref": "#/components/schemas/PeriodLimit"
            },
            "type": "array",
            "minItems": 1,
            "title": "Period Limits"
          },
          "allowed_merchant_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Allowed Merchant Ids"
          },
          "blocked_categories": {
            "items": {
              "type": "string",
              "enum": [
                "produce",
                "dairy",
                "eggs",
                "meat",
                "seafood",
                "bakery",
                "pantry",
                "beverage_non_alcoholic",
                "alcohol",
                "household",
                "unknown"
              ]
            },
            "type": "array",
            "title": "Blocked Categories"
          },
          "expires_at": {
            "type": "string",
            "title": "Expires At"
          },
          "approval_above_minor": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Approval Above Minor"
          },
          "velocity_limit": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VelocityLimit"
              },
              {
                "type": "null"
              }
            ]
          },
          "risk_review": {
            "type": "boolean",
            "title": "Risk Review",
            "default": false
          },
          "web_purchases": {
            "type": "boolean",
            "title": "Web Purchases",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "currency",
          "per_order_limit_minor",
          "period_limits",
          "allowed_merchant_ids",
          "blocked_categories",
          "expires_at",
          "approval_above_minor"
        ],
        "title": "Policy"
      },
      "DraftRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "text": {
            "type": "string",
            "minLength": 1
          },
          "delegatee_id": {
            "type": "string",
            "description": "Opaque server-generated identifier."
          },
          "parent_mandate_id": {
            "anyOf": [
              {
                "type": "string",
                "description": "Opaque server-generated identifier."
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "text",
          "delegatee_id"
        ]
      },
      "Ambiguity": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "field": {
            "type": "string"
          },
          "question": {
            "type": "string"
          }
        },
        "required": [
          "field",
          "question"
        ]
      },
      "DraftResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "draft_id": {
            "type": "string",
            "description": "Opaque server-generated identifier."
          },
          "owner_id": {
            "type": "string",
            "description": "Opaque server-generated identifier."
          },
          "delegatee_id": {
            "type": "string",
            "description": "Opaque server-generated identifier."
          },
          "parent_mandate_id": {
            "anyOf": [
              {
                "type": "string",
                "description": "Opaque server-generated identifier."
              },
              {
                "type": "null"
              }
            ]
          },
          "proposed_policy": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Policy"
              },
              {
                "type": "null"
              }
            ]
          },
          "ambiguities": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Ambiguity"
            }
          },
          "summary": {
            "type": "string"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "description": "RFC 3339 timestamp with an explicit UTC offset."
          }
        },
        "required": [
          "draft_id",
          "owner_id",
          "delegatee_id",
          "parent_mandate_id",
          "proposed_policy",
          "ambiguities",
          "summary",
          "expires_at"
        ]
      },
...
      "ShoppingItem": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "name": {
            "type": "string"
          },
          "quantity": {
            "type": "integer",
            "minimum": 1
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "name",
          "quantity"
        ]
      },
      "AgentRunRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "mandate_id": {
            "type": "string",
            "description": "Opaque server-generated identifier."
          },
          "shopping_list": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ShoppingItem"
            },
            "minItems": 1
          },
          "instruction": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "auto_purchase": {
            "type": "boolean",
            "default": false,
            "description": "Only within an already confirmed mandate. Default false for early integration."
          }
        },
        "required": [
          "mandate_id",
          "shopping_list"
        ]
      },
      "AgentRun": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "description": "Opaque server-generated identifier."
          },
          "mandate_id": {
            "type": "string",
            "description": "Opaque server-generated identifier."
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "awaiting_review",
              "quoted",
              "completed",
              "refused",
              "failed"
            ]
          },
          "provider": {
            "type": "string",
            "enum": [
              "jev",
              "fallback",
              "scripted"
            ]
          },
          "model_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "execution_mode": {
            "type": "string",
            "enum": [
              "local",
              "cloud",
              "scripted"
            ]
          },
          "quote_id": {
            "anyOf": [
              {
                "type": "string",
                "description": "Opaque server-generated identifier."
              },
              {
                "type": "null"
              }
            ]
          },
          "transaction_id": {
            "anyOf": [
              {
                "type": "string",
                "description": "Opaque server-generated identifier."
              },
              {
                "type": "null"
              }
            ]
          },
          "payment_id": {
            "anyOf": [
              {
                "type": "string",
                "description": "Opaque server-generated identifier."
              },
              {
                "type": "null"
              }
            ]
          },
          "latest_decision_id": {
            "anyOf": [
              {
                "type": "string",
                "description": "Opaque server-generated identifier."
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "RFC 3339 timestamp with an explicit UTC offset."
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "RFC 3339 timestamp with an explicit UTC offset."
          },
          "comparisons": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgentComparison"
            }
          }
        },
        "required": [
          "id",
          "mandate_id",
          "status",
          "provider",
          "model_id",
          "execution_mode",
          "quote_id",
          "transaction_id",
          "payment_id",
          "latest_decision_id",
          "message",
          "created_at",
          "updated_at"
        ]
      },
...
      "QuoteItem": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "product_id": {
            "type": "string",
            "description": "Opaque server-generated identifier."
          },
          "title": {
            "type": "string"
          },
          "quantity": {
            "type": "integer",
            "minimum": 1
          },
          "unit_price_minor": {
            "type": "integer",
            "minimum": 0
          },
          "line_total_minor": {
            "type": "integer",
            "minimum": 0
          },
          "category": {
            "$ref": "#/components/schemas/Category"
          },
          "category_status": {
            "type": "string",
            "enum": [
              "curated",
              "verified",
              "unknown",
              "conflicting"
            ]
          },
          "evidence_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "description": "Opaque server-generated identifier."
            }
          }
        },
        "required": [
          "product_id",
          "title",
          "quantity",
          "unit_price_minor",
          "line_total_minor",
          "category",
          "category_status",
          "evidence_ids"
        ]
      },
      "Charge": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "delivery",
              "other"
            ]
          },
          "label": {
            "type": "string"
          },
          "amount_minor": {
            "type": "integer",
            "minimum": 0
          },
          "evidence_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "description": "Opaque server-generated identifier."
            },
            "minItems": 1
          }
        },
        "required": [
          "kind",
          "label",
          "amount_minor",
          "evidence_ids"
        ]
      },
      "Quote": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "description": "Opaque server-generated identifier."
          },
          "merchant_id": {
            "type": "string",
            "description": "Opaque server-generated identifier."
          },
          "revision": {
            "type": "string"
          },
          "currency": {
            "$ref": "#/components/schemas/Currency"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuoteItem"
            }
          },
          "subtotal_minor": {
            "type": "integer",
            "minimum": 0
          },
          "charges": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Charge"
            }
          },
          "total_minor": {
            "type": "integer",
            "minimum": 0
          },
          "basket_hash": {
            "type": "string",
            "pattern": "^[a-f0-9]{64}$"
          },
          "delivery_context_id": {
            "type": "string",
            "description": "Opaque server-generated identifier."
          },
          "data_mode": {
            "type": "string",
            "enum": [
              "observed_snapshot",
              "observed_live"
            ]
          },
          "evidence_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "description": "Opaque server-generated identifier."
            }
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "RFC 3339 timestamp with an explicit UTC offset."
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "description": "RFC 3339 timestamp with an explicit UTC offset."
          }
        },
        "required": [
          "id",
          "merchant_id",
          "revision",
          "currency",
          "items",
          "subtotal_minor",
          "charges",
          "total_minor",
          "basket_hash",
          "delivery_context_id",
          "data_mode",
          "evidence_ids",
          "created_at",
          "expires_at"
        ],
        "description": "Authoritative for the sandbox merchant adapter, not proof of a live merchant checkout. Amounts and categories are computed server-side from catalog/evidence."
      },
...
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "error": {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "INVALID_REQUEST",
                  "UNAUTHENTICATED",
                  "FORBIDDEN",
                  "NOT_FOUND",
                  "IDEMPOTENCY_CONFLICT",
                  "STATE_CONFLICT",
                  "SERVICE_BUSY",
                  "MODEL_UNAVAILABLE",
                  "INTERNAL_ERROR"
                ]
              },
              "message": {
                "type": "string"
              },
              "request_id": {
                "type": "string"
              },
              "retryable": {
                "type": "boolean"
              },
              "details": {
                "type": "object",
                "additionalProperties": true
              }
            },
            "required": [
              "code",
              "message",
              "request_id",
              "retryable",
              "details"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
...
      "VelocityLimit": {
        "properties": {
          "max_purchases": {
            "type": "integer",
            "maximum": 1000.0,
            "minimum": 1.0,
            "title": "Max Purchases"
          },
          "window_minutes": {
            "type": "integer",
            "maximum": 10080.0,
            "minimum": 1.0,
            "title": "Window Minutes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "max_purchases",
          "window_minutes"
        ],
        "title": "VelocityLimit"
      },
...
      "AgentComparison": {
        "type": "object",
        "required": [
          "merchant_id",
          "quote",
          "matched_items",
          "missing_items",
          "problem"
        ],
        "properties": {
          "merchant_id": {
            "type": "string"
          },
          "quote": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Quote"
              },
              {
                "type": "null"
              }
            ]
          },
          "matched_items": {
            "type": "integer"
          },
          "missing_items": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "problem": {
            "type": "string"
          }
        }
      },
...
    }
  }
}